Explicitly control the generated node list
Through a combination of included/excluded nodes/reservations,
the hostlist to be tested can be eplicitly controlled.

The changes the default semantics. Before, `bench create` used
a hardcoded set of nodes to include. Now the host list to be
tested must be explicitly stated. Even checks for node state
have been removed.

Closes #26
Closes #25
Closes #22
Closes #1