First round of audits on bench.process
Untested: not guaranteed to work. Tests will come soon.

This consolidates a bunch of duplicated code, and makes
bench.process work a lot more like the new bench.add.

This also changes the paths that bench.process uses
to write out node lists, so bench.reserve will need
to be updated to match. (#59)