First pass at a clean-up of bench.reserve
Looks like pyslurm's reservation creation doesn't support
reservation naming; so we might want to go back to shelling
out until we can get that feature. (I'm already talking with
a pyslurm developer about it.)

Otherwise, I've rewritten bench.reserve to create
reservations for the bad_nodes list to be found in each
test directory. We should create a new flag (e.g., --not-tested)
that would create a reservation for nodes that are in
*all* not_tested files.