Remove hpl/scaling
The hpl tests, and the scaling test that referenced it, were
dependent on PBS, indicating that they haven't been used in
some time. In that light, I'm removing them. If we want an
hpl test in the future, it will probably be easier to
implement from scratch, rather than port this code.

Closes #64